Lazy recompute boundary when selection size has changed, DirtySelectionsResized, arap started
UpdateBoundary recomputed b if the selection has changed, returns true if b has changed

+++ Assets/Scripts/Libigl/LibiglBehaviour.Actions.cs
@@ -92,7 +92,7 @@ namespace Libigl
         {
             if (!ExecuteInput.DoHarmonicOnce) return;
             
-            Native.Harmonic(State, (int) ExecuteInput.VisibleSelectionMask, ExecuteInput.HarmonicShowDisplacement);
+            Native.Harmonic(State, ExecuteInput.VisibleSelectionMask, ExecuteInput.HarmonicShowDisplacement);
             State->DirtyState |= DirtyFlag.VDirty;
         }
